requirements.txt
# Core dependencies
psycopg2-binary>=2.9.5
lxml>=4.9.2
pandas>=1.5.2
numpy>=1.23.5
tqdm>=4.64.1

# Configuration and environment
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI
click>=8.1.3
colorama>=0.4.6
tabulate>=0.9.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
